untar: fix bugs in linkname and whiteout

+++ src/overlaybd/untar/whiteout.cpp
@@ -9,12 +9,10 @@
 #include <string>
 #include <set>
 #include <photon/fs/path.h>
-#include <photon/common/string_view.h>
 #include <photon/fs/filesystem.h>
 #include <photon/common/alog.h>
 #include <photon/common/alog-stdstring.h>
 #include <photon/common/enumerable.h>
-
 
 // whiteoutPrefix prefix means file is a whiteout. If this is followed by a
 // filename this means that file has been removed from the base layer.
@@ -26,27 +24,43 @@ const std::string whiteoutPrefix = ".wh.";
 // archives.
 const std::string whiteoutMetaPrefix = whiteoutPrefix + whiteoutPrefix;
 
-
 // whiteoutOpaqueDir file means directory has been made opaque - meaning
 // readdir calls to this directory do not follow to lower layers.
 const std::string whiteoutOpaqueDir = whiteoutMetaPrefix + ".opq";
 
 const std::string paxSchilyXattr = "SCHILY.xattr.";
 
+int Tar::mkdir_hier(const std::string_view &dir) {
+    struct stat s;
+    std::string path = remove_last_slash(dir);
+    if (fs->lstat(path.c_str(), &s) == 0) {
+        if (S_ISDIR(s.st_mode)) {
+            return 0;
+        } else {
+            errno = ENOTDIR;
+            // LOG_ERROR("mkdir ` failed, `", path.c_str(), strerror(errno));
+            return -1;
+        }
+    }
 
-int remove_all(photon::fs::IFileSystem *fs, const std::string &path) {
+    return photon::fs::mkdir_recursive(dir, fs, 0755);
+}
+
+int Tar::remove_all(const std::string &path, bool rmdir) {
     if (fs == nullptr || path.empty()) {
         LOG_ERROR("remove_all ` failed, fs is null or path is empty", path);
         return -1;
     }
     struct stat statBuf;
-    if (fs->lstat(path.c_str(), &statBuf) == 0) {        // get stat
+    if (fs->lstat(path.c_str(), &statBuf) == 0) { // get stat
         if (S_ISDIR(statBuf.st_mode) == 0) {      // not dir
-            fs->unlink(path.c_str());
+            if (unpackedPaths.find(path) == unpackedPaths.end()) {
+                fs->unlink(path.c_str());
+            }
             return 0;
         }
     } else {
-        LOG_DEBUG("get path ` stat failed, errno `:`", path, errno, strerror(errno));
+        LOG_ERROR("get path ` stat failed, errno `:`", path, errno, strerror(errno));
         return -1;
     }
 
@@ -58,13 +72,15 @@ int remove_all(photon::fs::IFileSystem *fs, const std::string &path) {
     dirent *dirInfo;
     while ((dirInfo = dirs->get()) != nullptr) {
         if (strcmp(dirInfo->d_name, ".") != 0 && strcmp(dirInfo->d_name, "..") != 0) {
-            remove_all(fs, path + "/" + dirInfo->d_name);
+            remove_all(path + "/" + dirInfo->d_name);
         }
         dirs->next();
     }
 
     fs->closedir(dirs);
-    fs->rmdir(path.c_str());
+    if (rmdir && unpackedPaths.find(path) == unpackedPaths.end()) {
+        fs->rmdir(path.c_str());
+    }
 
     return 0;
 }
@@ -78,23 +94,18 @@ int Tar::convert_whiteout(const char *filename) {
     auto base = p.basename();
     if (base == whiteoutOpaqueDir) {
         struct stat buf;
-        if (fs->lstat(dir.c_str(), &buf) < 0) {
+        std::string path = remove_last_slash(dir);
+        if (fs->lstat(path.c_str(), &buf) < 0) {
             LOG_ERRNO_RETURN(0, -1, "failed to lstat ", VALUE(dir));
         }
 
-        for (auto file : enumerable(photon::fs::Walker(fs, dir))) {
-            auto fn = std::string(file);
-            if (unpackedPaths.find(fn) == unpackedPaths.end()) {
-                remove_all(fs, fn);
-            }
-          }
+        remove_all(path, false);
         return 1;
     }
     if (base.substr(0, whiteoutPrefix.size()) == whiteoutPrefix) {
         auto opath = dir + std::string(base.data() + whiteoutPrefix.size());
-        remove_all(fs, opath);
+        remove_all(opath);
         return 1;
     }
     return 0;
 }
-
